# public/docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "FileFlow Project API NEW",
    "version": "1.0.0",
    "description": "API documentation for the FileFlow Project. All endpoints are synchronized with backend routes as of September 27, 2025."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/v1/admin/files": {
      "get": {
        "summary": "List All Files (Admin)",
        "description": "Returns a list of all files. Admin access required.",
        "responses": {
          "200": {
            "description": "List of files returned successfully."
          },
          "401": {
            "description": "Unauthorized."
          },
          "403": {
            "description": "Forbidden."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/v1/admin/files/{fileId}": {
      "get": {
        "summary": "Get File by ID (Admin)",
        "description": "Returns details for a specific file. Admin access required.",
        "parameters": [
          {
            "name": "fileId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "File details returned successfully."
          },
          "401": {
            "description": "Unauthorized."
          },
          "403": {
            "description": "Forbidden."
          },
          "404": {
            "description": "File not found."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      },
      "patch": {
        "summary": "Update File (Admin)",
        "description": "Updates a file's metadata. Admin access required.",
        "parameters": [
          {
            "name": "fileId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "File updated successfully."
          },
          "401": {
            "description": "Unauthorized."
          },
          "403": {
            "description": "Forbidden."
          },
          "404": {
            "description": "File not found."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      },
      "delete": {
        "summary": "Delete File (Admin)",
        "description": "Deletes a file. Admin access required.",
        "parameters": [
          {
            "name": "fileId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "File deleted successfully."
          },
          "401": {
            "description": "Unauthorized."
          },
          "403": {
            "description": "Forbidden."
          },
          "404": {
            "description": "File not found."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/v1/admin/files/bulk-delete": {
      "post": {
        "summary": "Bulk Delete Files (Admin)",
        "description": "Deletes multiple files. Admin access required.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Files deleted successfully."
          },
          "401": {
            "description": "Unauthorized."
          },
          "403": {
            "description": "Forbidden."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/v1/admin/files/stats/overview": {
      "get": {
        "summary": "Files Stats Overview (Admin)",
        "description": "Returns overview statistics for files. Admin access required.",
        "responses": {
          "200": {
            "description": "Stats overview returned successfully."
          },
          "401": {
            "description": "Unauthorized."
          },
          "403": {
            "description": "Forbidden."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    }
...
}
}